backend/requirements.txt
fastapi>=0.115.5
uvicorn[standard]>=0.32.1
python-dotenv>=1.0.1
python-multipart>=0.0.12
supabase>=2.10.0
anthropic>=0.40.0
PyMuPDF>=1.25.1
openpyxl>=3.1.5
stripe>=11.3.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.2.1
pydantic[email]>=2.11.0
httpx>=0.27.0
openai>=1.0.0
gspread>=6.0.0
google-auth>=2.0.0
gTTS>=2.5.0
